Filter Sending User Type

+++ moderation/server_settings/filters/send_word_blocker.py
@@ -6,33 +6,34 @@ from sigma.core.utilities.server_bound_logging import log_event
 
 async def send_word_blocker(ev, message):
     if message.guild:
-        prefix = ev.bot.get_prefix(message)
-        if not message.content.startswith(prefix):
-            text = clean_content(message.content.lower())
-            elements = text.split(' ')
-            blocked_words = ev.db.get_guild_settings(message.guild.id, 'BlockedWords')
-            if blocked_words is None:
-                blocked_words = []
-            remove = False
-            reason = None
-            for word in blocked_words:
-                if word in elements:
-                    remove = True
-                    reason = word
-                    break
-            if remove:
-                try:
-                    await message.delete()
-                    title = f'🔥 Your message was deleted for containing "{reason}".'
-                    to_author = discord.Embed(color=0xFFCC4D, title=title)
+        if type(message.author) == discord.Member:
+            prefix = ev.bot.get_prefix(message)
+            if not message.content.startswith(prefix):
+                text = clean_content(message.content.lower())
+                elements = text.split(' ')
+                blocked_words = ev.db.get_guild_settings(message.guild.id, 'BlockedWords')
+                if blocked_words is None:
+                    blocked_words = []
+                remove = False
+                reason = None
+                for word in blocked_words:
+                    if word in elements:
+                        remove = True
+                        reason = word
+                        break
+                if remove:
                     try:
-                        await message.author.send(embed=to_author)
-                    except discord.Forbidden:
+                        await message.delete()
+                        title = f'🔥 Your message was deleted for containing "{reason}".'
+                        to_author = discord.Embed(color=0xFFCC4D, title=title)
+                        try:
+                            await message.author.send(embed=to_author)
+                        except discord.Forbidden:
+                            pass
+                        author = f'{message.author.name}#{message.author.discriminator}'
+                        title = f'I deleted {author}\'s message for containing "{reason}".'
+                        log_embed = discord.Embed(color=0xFFCC4D)
+                        log_embed.set_author(name=title, icon_url=user_avatar(message.author))
+                        await log_event(ev.db, message.guild, log_embed)
+                    except discord.ClientException:
                         pass
-                    author = f'{message.author.name}#{message.author.discriminator}'
-                    title = f'I deleted {author}\'s message for containing "{reason}".'
-                    log_embed = discord.Embed(color=0xFFCC4D)
-                    log_embed.set_author(name=title, icon_url=user_avatar(message.author))
-                    await log_event(ev.db, message.guild, log_embed)
-                except discord.ClientException:
-                    pass
